more parts to parsing added
callbacks partially possible to parse. Still need more testing, though. Small adjustments made in ros2system to accommodate that. Also, easening the creation of Requests might be a matter of concern (maybe have factory-way of creating them similar to a client). Switched order of some of the data-fields for the sake of readability of output.